+++ examples/postEffect.js
@@ -40,7 +40,7 @@ for (const effect in effects) {
 let curEffect = 0
 const SPEED = 480
 
-gravity(2400)
+setGravity(2400)
 
 const level = addLevel([
 	// Design the level layout with symbols