+++ examples/runner.js
@@ -13,7 +13,7 @@ loadSprite("bean", "/sprites/bean.png")
 scene("game", () => {
 
 	// define gravity
-	gravity(2400)
+	setGravity(2400)
 
 	// add a game object to screen
 	const player = add([